Adds MIOpen to the project. (#27)
MIOpen carries a lot of deps and is a complicated project. This leaves a
couple of the optional deps off just to get things going.

Adds the following on the path to MIOpen:

* base/half
* math-libs/hipBLAS
* math-libs/rocRAND
* third-party/FunctionalPlus
* third-party/boost (just atomics, filesystem, system as static PIC
libraries)
* third-party/eigen
* third-party/frugally-deep
* third-party/nlohmann-json

Re-organized fetch_sources.py to allow controlling fetch of math-libs
and ml-frameworks.

Because we build with disaggregated libraries (vs glomming them all into
one prefix directory), this fuzzed what seems to be CMake config errors
in several of our libraries. I've patches around them vs fixing upstream
for the moment. See ml-frameworks/pre_hook_MIOpen.cmake for TODOs.