Add test project which finds and validates CMake packages and config. (#71)
Includes changes needed to get the packaging working and passing a test.

* Add a base/aux-overlay that populates some structural files that don't
belong to any project:
* CMake package redirection in the root lib/cmake directory to packages
in the lib/llvm/lib/cmake directory (so that users only need to specify
one CMAKE_PREFIX_PATH).
* An llvm -> lib/llvm symlink to account for an old directory
organization that is still required in some cases (should be removed
soon in favor of fixing upstream).
* Add `default_patterns` boolean to artifact component descriptors,
disabling the use of default includes/excludes for a component.
* Build `therock-dist(-.+)?` as part of ALL.
* Adds amd-comgr and hipcc to the amd-llvm artifact bundle (was an
oversight before).
* Includes `hipBLAS-common` in the `blas` artifact bundle (was an
oversight before).

Fixes #38