Patch rocprofiler-register for Windows compatibility and enable. (#94)
Progress on https://github.com/nod-ai/TheRock/issues/36.

This is gross, and the odd
[`.clang-format`](https://github.com/ROCm/rocprofiler-register/blob/amd-staging/.clang-format)
file in the rocprofiler-register repo isn't helping. There are probably
better design patterns for no-oping the entire project or adding proper
support for multiple platforms.

I hacked around enough to get a CMake configure + build that produces a
50KB `rocprofiler-register.dll` on Windows using MSVC and a default
CMake build on the default branch from GitHub (standalone, outside of
TheRock). The DLL is not expected to work since it is missing function
definitions and I did not implement Windows-specific versions of the
functions, but this unblocks more parts of build in this project that
depend on rocprofiler-register.

I did no testing on Linux, relying on our CI builds here for that.
Hopefully I didn't break anything.