Add logging and console control to the build. (#98)
* Adds a utility to handle console/log redirection and annotation.
* Adds an OUTPUT_ON_FAILURE flag for subprojects and sets it on all
third-party libs (some of which are very noisy), squelching all regular
console output unless if the step fails.
* Logs all project configure/build and timings to the logs/ directory.
* Enables an env var on the Linux build which uses GH
begingroup/endgroup folders. Leaves the Windows build as-is to compare
(but the new utility will prefix all log lines). Should decide which is
easier to consume.

The individual component builds are still far too noisy, but this gets
us back to manageable while also giving us log files to study more
closely. We should upload the logs directory as an artifact.