Enable HIPIFY on Windows. (#104)
Progress on https://github.com/nod-ai/TheRock/issues/36.

This code looks quite brittle:
https://github.com/ROCm/HIPIFY/blob/b803a5270bd9a63da25407181ca0346d472415e1/CMakeLists.txt#L90-L93.
I would expect any override of the compiler toolchain like that to be
performed via a different mechanism like a check and assert on the
existing value (not overwriting), a toolchain file,
`add_custom_command`, or `execute_process`... not overwriting the
compiler entirely like that.