Start portable linux packaging and isolate sysdeps. (#112)
* Adds a packaging/linux-portable/build_portable.py script which builds
the project under a manylinux docker container.
* Adds an option THEROCK_BUNDLE_SYSDEPS to enable system dependency
bundling (and enables by default on Linux).
* Adds a number of source built sysdeps for Linux:
  * bzip2
  * elfutils
  * libdrm
  * numactl
  * sqlite3
  * zlib
  * zstd
* Reworks project builds to depend on bundled sysdeps if available.
* All added source builds have been manipulated to prefix symbol version
and use prefixed/unique SONAMEs so that they can co-exist with any
system installed libraries in the same process.
* Inlines a patch to rocprofiler-sdk which aligns how it finds libdrm
with ROCR (it was just doing path munging).

The result is that for all current projects, no libraries part of the
ROCM distribution link to anything on the system other than
libc/libstdc++, which are generally considered to be well controlled and
always available.