Eliminate repo and switch to submodules. (#128)
I was down a dark path trying to get synchronized changes to pinned repo
manifests with the super-project (primarily because repo kind of
requires a secondary repository of manifests vs being able to use single
in-tree manifest files). The workflow is kind of the opposite of what we
want anyway (i.e. floating heads by default, very hard to pin in a
hermetic way).

This patch:

* Switches all repo project names to the same-named submodules.
* Eliminates the repo symlinks and just materializes the submodule at
the right place in the tree.
* Uses `update --recursive` to eliminate the ancillary fetches.
* Eliminates the Windows symlink hacks.
* Updates the patching/fetch_sources flow to work identically with
submodules.
* Commits the pins at the time I initialized.

We can make this nicer and eliminate fetch_sources entirely (or replace
it with some nicer helpers for doing partial submodule inits) once we
retire our local patches.

I will just manually `update --remote` daily until we get automation to
do it.