Add an initial test_linux_packages sub-workflow to our CI (#123)
This PR introduces testing on our MI300 fleet focused on installing
rocminfo and its dependencies that were built and uploaded into s3 from
the previous build_linux_packages phase.

A few callouts here:
1) Hard-coded dependencies for rocminfo, we should theoretically script
this out and be smarter about it.
2) Not running within a container on the MI300, ran into jank and didn't
seem worth it. The only thing I need from the build docker is aws so I
install it.
3) Have to be careful about extracting the tarball on /. Currently using rsync mechanism + fileset_tool will look into improvements in the future to not install in / but still ensure we are testing all the right libraries.

A few other notes, running this fast is a bit challenging. We may want
to add a couple other github actions to make debugging easier. What I
ended up doing was debugging with only the test part and a hardcoded
prior CI run on an ubuntu machine first. Switched to an MI300. Then
finally switched to depending on the build_linux_packages workflow. Will do this in follow-up PRs.